+++ scripts/mqtt-discovery-ble.shelly.js
@@ -1,6 +1,6 @@
 /**
  * BLE passive scanner and MQTT gateway
- * Version: 5.2.5
+ * Version: 5.2.6
  * Detected devices will be automatically registered to HA/Domoticz using MQTT Autodiscovery.
  *
  * Quick instructions: